+++ src/TodoList.jsx
@@ -1,0 +1,19 @@
+import React from 'react';
+
+const TodoList = () => {
+  const todoList = [
+    { id: 1, title: "Complete assignment" },
+    { id: 2, title: "Buy groceries" },
+    { id: 3, title: "Read a book" },
+  ];
+
+  return (
+    <ul>
+      {todoList.map(todo => (
+        <li key={todo.id}>{todo.title}</li>
+      ))}
+    </ul>
+  );
+};
+
+export default TodoList;
